move parallel_links_at_link, orientation_of_link into raster graph
