+++ landlab/graph/structured_quad/structured_quad.py
@@ -3,6 +3,7 @@ from functools import cached_property
 
 import numpy as np
 
+from ...grid.linkorientation import LinkOrientation
 from ...utils.decorators import read_only_array
 from ..graph import Graph
 
@@ -519,6 +520,102 @@ class StructuredQuadGraphExtras(StructuredQuadGraphTopology, Graph):
     @property
     def nodes_at_link(self):
         return self.ds["nodes_at_link"].values
+
+    @cached_property
+    def orientation_of_link(self):
+        """Return array of link orientation codes (one value per link).
+
+        Orientation codes are defined by :class:`~.LinkOrientation`;
+        1 = E, 2 = ENE, 4 = NNE, 8 = N, 16 = NNW, 32 = ESE (using powers
+        of 2 allows for future applications that might want additive
+        combinations).
+        """
+        orientation_of_link = np.full(
+            self.number_of_links, LinkOrientation.E, dtype=np.uint8
+        )
+        orientation_of_link[self.vertical_links] = LinkOrientation.N
+        return orientation_of_link
+
+    @cached_property
+    def parallel_links_at_link(self):
+        """Return similarly oriented links connected to each link.
+
+        Return IDs of links of the same orientation that are connected to
+        each given link's tail or head node.
+
+        The data structure is a *numpy* array of shape ``(n_links, 2)`` containing the
+        IDs of the "tail-wise" (connected to tail node) and "head-wise" (connected to
+        head node) links, or -1 if the link is inactive (e.g., on the perimeter) or
+        it has no attached parallel neighbor in the given direction.
+
+        For instance, consider a 3x4 raster, in which link IDs are as shown::
+
+            .-14-.-15-.-16-.
+            |    |    |    |
+            10  11   12   13
+            |    |    |    |
+            .--7-.--8-.--9-.
+            |    |    |    |
+            3    4    5    6
+            |    |    |    |
+            .--0-.--1-.--2-.
+
+        Here's a mapping of the tail-wise (shown at left or bottom of links) and
+        head-wise (shown at right or top of links) links::
+
+            .----.----.----.
+            |    |    |    |
+            |    |    |    |
+            |    4    5    |
+            .---8.7--9.8---.
+            |   11   12    |
+            |    |    |    |
+            |    |    |    |
+            .----.----.----.
+
+        So the corresponding data structure would be mostly filled with -1, but
+        for the 7 active links, it would look like::
+
+            4: [[-1, 11],
+            5:  [-1, 12],
+            7:  [-1,  8],
+            8:  [ 7,  9],
+            9:  [ 8, -1],
+            11: [ 4, -1],
+            12: [ 5, -1]]
+
+        Examples
+        --------
+        >>> from landlab import RasterModelGrid
+        >>> grid = RasterModelGrid((3, 4))
+        >>> pll = grid.parallel_links_at_link
+        >>> pll[4:13, :]
+        array([[-1, 11],
+               [-1, 12],
+               [-1, -1],
+               [-1,  8],
+               [ 7,  9],
+               [ 8, -1],
+               [-1, -1],
+               [ 4, -1],
+               [ 5, -1]])
+        """
+        plinks_at_link = np.full((self.number_of_links, 2), -1, dtype=int)
+
+        plinks_at_link[self.vertical_links, 0] = self.links_at_node[
+            self.node_at_link_tail[self.vertical_links], 3
+        ]
+        plinks_at_link[self.vertical_links, 1] = self.links_at_node[
+            self.node_at_link_head[self.vertical_links], 1
+        ]
+        plinks_at_link[self.horizontal_links, 0] = self.links_at_node[
+            self.node_at_link_tail[self.horizontal_links], 2
+        ]
+        plinks_at_link[self.horizontal_links, 1] = self.links_at_node[
+            self.node_at_link_head[self.horizontal_links], 0
+        ]
+
+        return plinks_at_link
 
 
 class StructuredQuadGraph(StructuredQuadGraphExtras):
