+++ landlab/grid/raster.py
@@ -387,108 +387,6 @@ class RasterModelGrid(DiagonalsMixIn, DualUniformRectilinearGraph, ModelGrid):
         """
         return (self.number_of_cell_rows, self.number_of_cell_columns)
 
-    @property
-    def parallel_links_at_link(self):
-        """Return similarly oriented links connected to each link.
-
-        Return IDs of links of the same orientation that are connected to
-        each given link's tail or head node.
-
-        The data structure is a *numpy* array of shape ``(n_links, 2)`` containing the
-        IDs of the "tail-wise" (connected to tail node) and "head-wise" (connected to
-        head node) links, or -1 if the link is inactive (e.g., on the perimeter) or
-        it has no attached parallel neighbor in the given direction.
-
-        For instance, consider a 3x4 raster, in which link IDs are as shown::
-
-            .-14-.-15-.-16-.
-            |    |    |    |
-            10  11   12   13
-            |    |    |    |
-            .--7-.--8-.--9-.
-            |    |    |    |
-            3    4    5    6
-            |    |    |    |
-            .--0-.--1-.--2-.
-
-        Here's a mapping of the tail-wise (shown at left or bottom of links) and
-        head-wise (shown at right or top of links) links::
-
-            .----.----.----.
-            |    |    |    |
-            |    |    |    |
-            |    4    5    |
-            .---8.7--9.8---.
-            |   11   12    |
-            |    |    |    |
-            |    |    |    |
-            .----.----.----.
-
-        So the corresponding data structure would be mostly filled with -1, but
-        for the 7 active links, it would look like::
-
-            4: [[-1, 11],
-            5:  [-1, 12],
-            7:  [-1,  8],
-            8:  [ 7,  9],
-            9:  [ 8, -1],
-            11: [ 4, -1],
-            12: [ 5, -1]]
-
-        Examples
-        --------
-        >>> from landlab import RasterModelGrid
-        >>> grid = RasterModelGrid((3, 4))
-        >>> pll = grid.parallel_links_at_link
-        >>> pll[4:13, :]
-        array([[-1, 11],
-               [-1, 12],
-               [-1, -1],
-               [-1,  8],
-               [ 7,  9],
-               [ 8, -1],
-               [-1, -1],
-               [ 4, -1],
-               [ 5, -1]])
-        """
-        try:
-            pll = self._parallel_links_at_link
-        except AttributeError:
-            pll = self._setup_parallel_links_at_link()
-        return pll
-
-    def _setup_parallel_links_at_link(self):
-        """
-        For a regular grid, set up a data structure mapping each link to its parallel
-        and connected neighbors.
-
-        (For example and doctest, see `parallel_links_at_link`)
-        """
-        plinks_at_link = -np.ones((self.number_of_links, 2), dtype=int)
-        plinks_at_link[self.vertical_links, 0] = self.links_at_node[
-            self.node_at_link_tail[self.vertical_links], 3
-        ]
-        plinks_at_link[self.vertical_links, 1] = self.links_at_node[
-            self.node_at_link_head[self.vertical_links], 1
-        ]
-        plinks_at_link[self.horizontal_links, 0] = self.links_at_node[
-            self.node_at_link_tail[self.horizontal_links], 2
-        ]
-        plinks_at_link[self.horizontal_links, 1] = self.links_at_node[
-            self.node_at_link_head[self.horizontal_links], 0
-        ]
-        boundary_tail = (
-            self.status_at_node[self.node_at_link_tail] != self.BC_NODE_IS_CORE
-        )
-        plinks_at_link[boundary_tail, 0] = -1
-        boundary_head = (
-            self.status_at_node[self.node_at_link_head] != self.BC_NODE_IS_CORE
-        )
-        plinks_at_link[boundary_head, 1] = -1
-
-        self._parallel_links_at_link = plinks_at_link
-        return self._parallel_links_at_link
-
     def _create_link_unit_vectors(self):
         """Make arrays to store the unit vectors associated with each link.
